requirements.txt
Flask==2.1.2
requests~=2.28.1
beautifulsoup4~=4.11.1
python-dateutil~=2.8.2